Add lambda parsing
